+++ src/main/java/racingcar/exception/ExceptionMessage.java
@@ -1,13 +1,13 @@
-package racingcar.error;
+package racingcar.exception;
 
-public enum ErrorMessage {
+public enum ExceptionMessage {
     INVALID_NAME_LENGTH("이름은 %d글자 이하여야 합니다."),
     INVALID_COUNT("시도 횟수는 %d회 이상이여야 합니다.");
 
     private final String value;
     private static final String ERROR_HEAD = "[ERROR]: %s";
 
-    ErrorMessage(String message) {
+    ExceptionMessage(String message) {
         this.value = String.format(ERROR_HEAD, message);
     }
